payments/serializers.py
from rest_framework import serializers
from .models import PaymentTransaction


class PaymentTransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = PaymentTransaction
        fields = ("id", "phone_number", "amount", "status", "mpesa_checkout_request_id", "created_at")

